TrashDetector/requirements.txt
pandas
numpy
matplotlib
streamlit
ultralytics
zenml[server]
mlflow
pillow